+++ metreeca-text/src/main/java/com/metreeca/text/actions/Normalize.java
@@ -21,11 +21,44 @@ import java.text.Normalizer.Form;
 import java.util.function.UnaryOperator;
 
 /**
- * Text cleaning.
+ * Text normalization.
  *
- * <p>Cleans and normalizes text values.</p>
+ * <p>Normalizes and cleans text values.</p>
  */
 public final class Normalize implements UnaryOperator<String> {
+
+	private static final Normalize Default=new Normalize()
+			.space(true)
+			.smart(true)
+			.marks(true);
+
+
+	/**
+	 * Normalizes a text value.
+	 *
+	 * @param text the text value to be normalized
+	 *
+	 * @return a version of {@code text} normalized with the following parameters:
+	 *
+	 * <ul>
+	 *     <li>{@link #space(boolean) space(true)}</li>
+	 *     <li>{@link #smart(boolean) smart(true)}</li>
+	 *     <li>{@link #marks(boolean) marks(true)}</li>
+	 * </ul>
+	 *
+	 * @throws NullPointerException if {@code text} is null
+	 */
+	public static String normalize(final String text) {
+
+		if ( text == null ) {
+			throw new NullPointerException("null text");
+		}
+
+		return Default.apply(text);
+	}
+
+
+	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 	private boolean space;
 	private boolean marks;